@@ -5,52 +5,51 @@ using Meadow.Units;
 using System;
 using System.Threading.Tasks;
 
-namespace AnalogInputPort
+namespace AnalogInputPort;
+
+public class MeadowApp : App<F7FeatherV2>
 {
-    public class MeadowApp : App<F7FeatherV2>
+    IAnalogInputPort analogIn;
+
+    public override Task Initialize()
     {
-        IAnalogInputPort analogIn;
+        Resolver.Log.Info("Initializing hardware...");
 
-        public override Task Initialize()
+        analogIn = Device.CreateAnalogInputPort(Device.Pins.A00);
+
+        analogIn.Updated += (s, result) =>
         {
-            Resolver.Log.Info("Initializing hardware...");
+            Resolver.Log.Info($"Analog event, new voltage: {result.New.Volts:N2}V, old: {result.Old?.Volts:N2}V");
+        };
 
-            analogIn = Device.CreateAnalogInputPort(Device.Pins.A00);
-
-            analogIn.Updated += (s, result) =>
+        var observer = IAnalogInputPort.CreateObserver(
+            handler: result =>
             {
-                Resolver.Log.Info($"Analog event, new voltage: {result.New.Volts:N2}V, old: {result.Old?.Volts:N2}V");
-            };
-
-            var observer = IAnalogInputPort.CreateObserver(
-                handler: result =>
+                Resolver.Log.Info($"Analog observer triggered; new: {result.New.Volts:n2}V, old: {result.Old?.Volts:n2}V");
+            },
+            // filter is optional. in this case, we're only notifying if the
+            // voltage changes by at least `0.1V`.
+            filter: result =>
+            {
+                if (result.Old is { } oldValue)
                 {
-                    Resolver.Log.Info($"Analog observer triggered; new: {result.New.Volts:n2}V, old: {result.Old?.Volts:n2}V");
-                },
-                // filter is optional. in this case, we're only notifying if the
-                // voltage changes by at least `0.1V`.
-                filter: result =>
-                {
-                    if (result.Old is { } oldValue)
-                    {
-                        return (result.New - oldValue).Abs().Volts > 0.1;
-                    }
-                    else { return false; }
+                    return (result.New - oldValue).Abs().Volts > 0.1;
                 }
-            );
-            analogIn.Subscribe(observer);
+                else { return false; }
+            }
+        );
+        analogIn.Subscribe(observer);
 
-            Resolver.Log.Info("Hardware initialized.");
+        Resolver.Log.Info("Hardware initialized.");
 
-            return base.Initialize();
-        }
+        return base.Initialize();
+    }
 
-        public override async Task Run()
-        {
-            Voltage voltageReading = await analogIn.Read();
-            Resolver.Log.Info($"Voltages: {voltageReading.Volts:N3}");
+    public override async Task Run()
+    {
+        Voltage voltageReading = await analogIn.Read();
+        Resolver.Log.Info($"Voltages: {voltageReading.Volts:N3}");
 
-            analogIn.StartUpdating(TimeSpan.FromSeconds(1));
-        }
+        analogIn.StartUpdating(TimeSpan.FromSeconds(1));
     }
 }
