@@ -1,21 +1,19 @@
 ﻿using Meadow;
 using Meadow.Devices;
-using System;
 using System.Threading.Tasks;
 
-namespace BatteryLevel
+namespace BatteryLevel;
+
+public class MeadowApp : App<F7FeatherV2>
 {
-    public class MeadowApp : App<F7FeatherV2>
+    public override async Task Run()
     {
-        public override async Task Run()
+        while (true)
         {
-            while (true)
-            {
 
-                Resolver.Log.Info($"Battery Voltage: {Device.GetBatteryInfo().Voltage.Value.Volts:N2}V");
+            Resolver.Log.Info($"Battery Voltage: {Device.GetBatteryInfo().Voltage.Value.Volts:N2}V");
 
-                await Task.Delay(3000);
-            }
+            await Task.Delay(3000);
         }
     }
 }
