@@ -1,20 +1,19 @@
 ﻿using Meadow;
 
-namespace Threading_Basics
+namespace ResolverServices;
+
+public class InjectedConstructorService : IService
 {
-    public class InjectedConstructorService : IService
+    private IOutputService OutputService { get; }
+
+    public void SetOutputState(bool state)
     {
-        private IOutputService OutputService { get; }
+        OutputService.OutputPort.State = state;
+    }
 
-        public void SetOutputState(bool state)
-        {
-            OutputService.OutputPort.State = state;
-        }
-
-        public InjectedConstructorService(IOutputService outputService)
-        {
-            Resolver.Log.Info($"InjectedConstructorService constructor has been called.");
-            OutputService = outputService;
-        }
+    public InjectedConstructorService(IOutputService outputService)
+    {
+        Resolver.Log.Info($"InjectedConstructorService constructor has been called.");
+        OutputService = outputService;
     }
 }
