@@ -1,14 +1,13 @@
-﻿using System;
-using SQLite;
+﻿using SQLite;
+using System;
 
-namespace SQLite_Sample
+namespace SQLite_Sample;
+
+[Table("SensorReadings")]
+public class SensorModel
 {
-    [Table("SensorReadings")]
-    public class SensorModel
-    {
-        [PrimaryKey, AutoIncrement]
-        public int ID { get; set; }
-        public DateTime Timestamp { get; set; }
-        public double Value { get; set; }
-    }
+    [PrimaryKey, AutoIncrement]
+    public int ID { get; set; }
+    public DateTime Timestamp { get; set; }
+    public double Value { get; set; }
 }
