@@ -1,19 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace HttpListener_Basics;
 
-namespace HttpListener_Basics
+public class Secrets
 {
-    public class Secrets
-    {
-        /// <summary>
-        /// WiFi name or SSID
-        /// </summary>
-        public const string WIFI_NAME = "WIFI_NAME";
+    /// <summary>
+    /// WiFi name or SSID
+    /// </summary>
+    public const string WIFI_NAME = "WIFI_NAME";
 
-        /// <summary>
-        /// WiFi password.
-        /// </summary>
-        public const string WIFI_PASSWORD = "WIFI_PASSWORD";
-    }
+    /// <summary>
+    /// WiFi password.
+    /// </summary>
+    public const string WIFI_PASSWORD = "WIFI_PASSWORD";
 }
